ref!: swap secp256k1 libs and dashkeys for base58check
