+++ package.json
@@ -7,7 +7,6 @@
     "dashwallet": "bin/wallet.js"
   },
   "files": [
-    "dashapi.js",
     "index.js",
     "lib/",
     "wallet.js"
@@ -41,12 +40,9 @@
   "dependencies": {
     "@dashevo/dashcore-lib": "^0.19.44",
     "@dashincubator/secp256k1": "^1.7.1-5",
-    "@root/base58check": "^1.0.1",
     "dashhd": "^3.0.5",
     "dashphrase": "^1.3.9",
     "dashsight": "^1.5.0",
-    "dotenv": "^16.0.3",
-    "ripemd160": "^2.0.2",
-    "secp256k1": "^4.0.3"
+    "dotenv": "^16.0.3"
   }
 }
